ReaperAutomation/requirements.txt
psutil==5.9.5
python-reapy==0.10.0
typing_extensions==4.6.3
